# [Thinkit] Replaced GetGenericTestbed with GetTestbedWithRequirements to take in TestbedRequiredments proto, Fix PacketGenerationFinalizer and RequestType enum, Benchmark performance of Bazel Test Environment, Add a way to set test_case_id with Bazel Test Environment, Add a mutex to BazelTestEnvironment making it thread safe, Language change to remove disrespectful term&  Add a test for BazelTestEnvironment using protos.