--- a/thinkit/BUILD.bazel
+++ b/thinkit/BUILD.bazel
@@ -170,6 +170,7 @@ cc_library(
         "//gutil:status",
         "@com_google_absl//absl/status",
         "@com_google_absl//absl/strings",
+        "@com_google_absl//absl/synchronization",
         "@com_google_googletest//:gtest",
     ],
 )
@@ -179,9 +180,16 @@ cc_test(
     srcs = ["bazel_test_environment_test.cc"],
     deps = [
         ":bazel_test_environment",
-        "//gutil:status_matchers",
-        "@com_google_absl//absl/strings",
+        ":test_environment",
+        "@com_google_protobuf//:protobuf",
         "@com_google_googletest//:gtest_main",
+        "@com_google_absl//absl/memory",
+        "@com_google_absl//absl/strings",
+        # This dependency is not technically needed inside Google3,
+        # but is a hack to keep the Copybara rules simple.
+        "@com_google_benchmark//:benchmark",
+        "//gutil:status_matchers",
+        "//gutil:testing",
     ],
 )
 
@@ -239,7 +247,9 @@ cc_library(
     hdrs = ["generic_testbed_fixture.h"],
     deps = [
         ":generic_testbed",
+        "//thinkit/proto:generic_testbed_cc_proto",
         "@com_google_absl//absl/memory",
+        "@com_google_absl//absl/status:statusor",
         "@com_google_googletest//:gtest",
     ],
 )