--- a/thinkit/packet_generation_finalizer.h
+++ b/thinkit/packet_generation_finalizer.h
@@ -25,9 +25,7 @@ class PacketGenerationFinalizer {
   virtual ~PacketGenerationFinalizer() = 0;
 };
 
-
 inline PacketGenerationFinalizer::~PacketGenerationFinalizer() {}
-
 
 }  // namespace thinkit
 #endif  // GOOGLE_THINKIT_PACKET_GENERATION_FINALIZER_H_